Code refactored (It's less shit!)
So I looked back at this project and had a bit of a cringe moment
scrolling through the code, so I did some fixes and changed. I expect to
repeat this cycle in about 5-6 months, but for now it works. It was
never broken, and not a lot was fixed - I just made it look and work
nicer.