Gui removed and JsonHelper updated!
Since I can't be arsed to work on the GUI version because I don't find
it fun enough, I've removed it from the project. However I do have it
backed up incase an intrest would spark in the future. JsonHelper was
also updated and is now included in the HourBoostr release.